+++ src/client/vite-env.d.ts
@@ -1,8 +1,3 @@
 /// <reference types="vite/client" />
 
 declare const __APP_GIT_SHA__: string;
-
-declare module "guacamole-common-js" {
-  const Guacamole: any;
-  export default Guacamole;
-}